Simplify API endpoints to basic mock responses for easier development
